Guard and reorganize actix-specific scabbard code
Moves the actix-specific code for the scabbard REST API to a new `actix`
submodule and guards it with the `rest-api-actix` feature flag. Also
separates the endpoints into their own submodules. This matches the
pattern used for other REST API endpoints in splinter.
